Move workflow reconnection utility
